Added a FieldMixin and removed required parameter in Form Fields
Also renamed setValue, initValue to setField, initField.

The FieldMixin wraps the context in methods of their own to hide how everything works, and provide an easier interface to work with. It also is able to provide some defaults like name, value, etc when calling these functions. Defaults can all be overriden by just passing them into the object parameter.

Required didn't do much of anything. All it did was disable the validator from stopping the form from being submitted. Now, if a validator exists, the field is "required."However, it is still used in the fields to render the `*`.